Moved loadSetting to MyAvatar
